// gravitee-apim-portal-webui/projects/portal-webclient-sdk/src/lib/model/finalizeRegistrationInput.ts
/**
 * Gravitee.io Portal Rest API
 * API dedicated to the devportal part of Gravitee
 *
 * The version of the OpenAPI document: 3.12.0-SNAPSHOT
 * Contact: contact@graviteesource.com
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export interface FinalizeRegistrationInput { 
    /**
     * Token of the registered user to be validated.
     */
    token: string;
    /**
     * Password of the registered user.
     */
    password: string;
    /**
     * First name of the registered user.
     */
    firstname: string;
    /**
     * Last name of the registered user.
     */
    lastname: string;
}